+++ src/actions/CheckAccessTrait.php
@@ -8,14 +8,15 @@
 
 namespace flipbox\craft\ember\actions;
 
+use yii\web\ForbiddenHttpException;
 use yii\web\UnauthorizedHttpException;
 
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
  * @since 2.0.0
  *
- * @property int $statusCodeUnauthorized
- * @property string $messageUnauthorized
+ * @property int $errorCodeForbidden
+ * @property string $messageForbidden
  */
 trait CheckAccessTrait
 {
@@ -27,14 +28,14 @@ trait CheckAccessTrait
     /**
      * @param mixed ...$params
      * @return mixed
-     * @throws UnauthorizedHttpException
+     * @throws ForbiddenHttpException
      */
     public function checkAccess(...$params)
     {
         if ($this->checkAccess) {
             if (call_user_func_array($this->checkAccess, $params) === false) {
                 /** @noinspection PhpVoidFunctionResultUsedInspection */
-                return $this->handleUnauthorizedResponse();
+                return $this->handleForbiddenResponse();
             };
         }
 
@@ -45,6 +46,8 @@ trait CheckAccessTrait
      * HTTP forbidden response code
      *
      * @return int
+     *
+     * @deprecated
      */
     protected function statusCodeUnauthorized(): int
     {
@@ -53,6 +56,8 @@ trait CheckAccessTrait
 
     /**
      * @return string
+     *
+     * @deprecated
      */
     protected function messageUnauthorized(): string
     {
@@ -61,12 +66,43 @@ trait CheckAccessTrait
 
     /**
      * @throws UnauthorizedHttpException
+     *
+     * @deprecated
      */
     protected function handleUnauthorizedResponse()
     {
         throw new UnauthorizedHttpException(
             $this->messageUnauthorized(),
             $this->statusCodeUnauthorized()
+        );
+    }
+
+    /**
+     * HTTP forbidden response code
+     *
+     * @return int|null
+     */
+    protected function errorCodeForbidden()
+    {
+        return $this->errorCodeForbidden;
+    }
+
+    /**
+     * @return string
+     */
+    protected function messageForbidden(): string
+    {
+        return $this->messageForbidden ?? 'Unable to perform action.';
+    }
+
+    /**
+     * @throws ForbiddenHttpException
+     */
+    protected function handleForbiddenResponse()
+    {
+        throw new ForbiddenHttpException(
+            $this->messageForbidden(),
+            $this->errorCodeForbidden()
         );
     }
 }