+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+
+## Unreleased
+### Fixed
+- Action::checkAccess now throws a ForbiddenHttpException instead of a UnauthorizedHttpException.
+
 ## 2.5.3 - 2019-10-30
 ### Fixed
 - When passing an empty array to `Record::findAll([])` an improper condition would be built.